% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annualize_time_series.R
\name{annualize_time_series}
\alias{annualize_time_series}
\title{Transform the quarterly landsat data to an annual time series}
\usage{
annualize_time_series(
  segmented_landsat_data,
  ...,
  annualization_method = c("max_first", "sum_first")
)
}
\arguments{
\item{segmented_landsat_data}{dataframe generated by \code{\link[=segment_landsat_data]{segment_landsat_data()}}
with rows for each pixel, and columns for each quarter and year of the
landsat data, with the assigned segment from a given shapefile}

\item{...}{unused}

\item{annualization_method}{The method used to convert the quarterly time series to an annual time
series.
\itemize{
\item "max_first": for each pixel, the maximum quarterly area in a year is calculated first, and
then all these maximum areas per year are summed together within each kelp segment.
\item "sum_first": sums together all quarterly area values for each pixel within each segment, and
then the maximum area per year
for each segment is calculated
}}
}
\value{
A dataframe with the following columns:
\itemize{
\item \code{Segment_ID}: The ID of the kelp area segment
\item \code{max_occupiable}: The maximum occupiable kelp area for the given segment. This column repeats
the same value for all rows in a single kelp area segment
\item \code{historical_med}: The historical (1984-2013) median kelp area for the segment. The median is
computed across the maximum yearly kelp area. This column repeats the same value for all rows
in a single kelp area segment
\item \code{year}: year of the time series
\item \code{area_abs}: kelp area in absolute magnitude, in km^2
\item \code{area_hist}: kelp area relative to the historical median, expressed as a percentage
\item \code{area_pct}: kelp area relative to the maximum occupiable kelp area, expressed as a percentage
}
}
\description{
The landsat kelp area data is a quarterly product, so this function
serves to transform it to an annual product, with two separate options, described below.
}
\seealso{
\code{\link[=get_max_occupiable_area]{get_max_occupiable_area()}}
}
