% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_time_series.R
\name{extract_time_series}
\alias{extract_time_series}
\title{Transform the quarterly landsat data to an annual time series}
\usage{
extract_time_series(
  segmented_landsat_data,
  ...,
  frequency = c("annual", "quarterly"),
  annualization_method = c("sum_first", "max_first", "Q1", "Q2", "Q3", "Q4"),
  segment_id = NULL
)
}
\arguments{
\item{segmented_landsat_data}{A dataframe generated by \code{\link[=segment_landsat_data]{segment_landsat_data()}}
with rows for each pixel, and columns for each quarter and year of the
landsat data, with the assigned segment from a given shapefile}

\item{...}{These dots are for future extensions and must be empty.}

\item{frequency}{The frequency of the time series to extract. Either "annual" (default) or "quarterly"}

\item{annualization_method}{The method used to convert the quarterly time series to an annual time
series.
\itemize{
\item "sum_first": sums together all quarterly area values for each pixel within each segment, and
then the maximum area per year for each segment is calculated
\item "max_first": for each pixel, the maximum quarterly area in a year is calculated first, and
then all these maximum areas per year are summed together within each kelp segment.
\item "Q1": selects the first quarter of every year
\item "Q2": selects the second quarter of every year
\item "Q3": selects the third quarter of every year
\item "Q4": selects the fourth quarter of every year
}}

\item{segment_id}{A character vector containing the subset of \code{segment_id}'s of interest, e.g.
\code{c("CA_71", "CA_78")}, or \code{NULL} for all segment IDs (default).}
}
\value{
A dataframe with the following columns:
\itemize{
\item \code{Segment_ID}: The ID of the kelp area segment
\item \code{max_occupiable}: The maximum occupiable kelp area for the given segment. This column repeats
the same value for all rows in a single kelp area segment
\item \code{historical_med}: The historical (1984-2013) median kelp area for the segment. The median is
computed across the annualized kelp area values. This column repeats the same value for all rows
in a single kelp area segment
\item \code{quarter}*: If \code{frequency == "quarterly"}, the quarter in the time series, otherwise not present
\item \code{year}: Year in the time series
\item \code{date}: date representation of the year/quarter, for simpler plotting
\item \code{area_abs}: Kelp area in absolute magnitude, in km^2
\item \code{area_hist}: Kelp area relative to the historical median, expressed as a percentage
\item \code{area_pct}: Kelp area relative to the maximum occupiable kelp area, expressed as a percentage
}
}
\description{
The Landsat kelp area data is a quarterly product per pixel, so this function
serves to aggregate these pixels to a time series for each segment, either quarterly or
annually, with a few different methods for the annualization.
}
\seealso{
\code{\link[=get_max_occupiable_area]{get_max_occupiable_area()}}
}
