% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_kelp_presence.R
\name{get_kelp_presence}
\alias{get_kelp_presence}
\title{Get kelp presence within each segment}
\usage{
get_kelp_presence(
  segmented_landsat_data,
  kelp_segments_file_path,
  ...,
  no_kelp_bound = 0.02,
  ephemeral_kelp_bound = 0.15
)
}
\arguments{
\item{segmented_landsat_data}{A dataframe generated by \code{\link[=segment_landsat_data]{segment_landsat_data()}}
with rows for each pixel, and columns for each quarter and year of the
landsat data, with the assigned segment from a given shapefile}

\item{kelp_segments_file_path}{character string path to the kelp segments shapefile. Must contain
a Segment_ID attribute and use the WGS 84 coordinate reference system (EPSG:4326)}

\item{...}{These dots are for future extensions and must be empty.}

\item{no_kelp_bound}{Upper bound, in percent of segment area, for the "No Historical Kelp"
category. Default is 0.02}

\item{ephemeral_kelp_bound}{Upper bound, in percentage of segment area, for the "Ephemeral
Kelp" category. Default is 0.15}
}
\value{
A dataframe with the following columns:
\itemize{
\item \code{Segment_ID}: The ID of the kelp area segment
\item \code{segment_area}: The area in km2 of the kelp segment
\item \code{pixel_percent}: The percentage of the kelp segment's area that the total number of pixels
ever detected occupy. This calculation assumes each pixel contains the full 900 m2 of kelp
area
\item \code{presence}: The category of kelp presence for each segment.
\itemize{
\item "No Historical Kelp": \code{pixel_percent} <= \code{no_kelp_bound} OR \code{pixel_percent} is \code{NA}
\item "Ephemeral Kelp": \code{pixel_percent} <= \code{ephemeral_kelp_bound}
\item "Kelp": otherwise
}
}
}
\description{
Many kelp segments may have only a few pixels ever detected within them, so this function serves
to classify these different kelp segments, based on the area that each detected pixel occupies
relative to the area of the entire kelp segment.
}
