% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_kelp_status.R
\name{get_kelp_status}
\alias{get_kelp_status}
\title{Get the kelp status for a particular year}
\usage{
get_kelp_status(
  annual_time_series,
  kelp_presence,
  ...,
  status_year = NULL,
  segment_id = NULL
)
}
\arguments{
\item{annual_time_series}{Annual time series generated by \code{\link[=extract_time_series]{extract_time_series()}}.
Does not work with quarterly time series}

\item{kelp_presence}{Kelp presence data frame generated by \code{\link[=get_kelp_presence]{get_kelp_presence()}}}

\item{...}{unused}

\item{status_year}{Year to calculate the status for, or \code{NULL} for the latest year in the time series.
Must be later than 2013}

\item{segment_id}{A character vector containing the subset of \code{segment_id}'s of interest, e.g.
\code{c("CA_71", "CA_78")}, or \code{NULL} for all segment IDs (default).}
}
\value{
A data frame with statuses for each Segment_ID in \code{annual_time_series}. A value of -999
indicates that there is not enough data to determine a status for the given kelp segment
}
\description{
Status is calculated by dividing the given year's kelp area by the historical median across 1984-2013
for each kelp segment.
}
