% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_max_occupiable_area.R
\name{get_max_occupiable_kelp_area}
\alias{get_max_occupiable_kelp_area}
\title{Calculate the maximum occupiable kelp area}
\usage{
get_max_occupiable_kelp_area(segmented_landsat_data)
}
\arguments{
\item{segmented_landsat_data}{A dataframe generated by \code{KelpAreaIndicator::segment_landsat_data()}
with rows for each pixel, and columns for each quarter/year combination of the
landsat data, with the assigned segment from a given shapefile in the \code{Segment_ID} column}
}
\value{
A dataframe with the maximum occupiable area for each kelp segment
}
\description{
Maximum occupiable kelp area is the maximum possible area that kelp has ever
occupied in the landsat area data, e.g. if a pixel showed 0.003 km2 kelp coverage once in the
40 year time series, and 0 km2 the remainder of the time, the maximum occupiable
kelp area for that pixel is 0.003 km2. For an entire segment, the maximum occupiable kelp area
is the sum of each of its pixel's maximum occupiable kelp area.
}
