% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_recent_trend.R
\name{get_recent_trend}
\alias{get_recent_trend}
\title{Calculate recent trends from annual kelp area time series}
\usage{
get_recent_trend(
  annual_time_series,
  ...,
  ref_year = NULL,
  num_recent_years = 5
)
}
\arguments{
\item{annual_time_series}{Annual time series generated by \code{\link[=extract_time_series]{extract_time_series()}}.
Does not work with quarterly time series}

\item{...}{These dots are for future extensions and must be empty.}

\item{ref_year}{Reference year or years from which to calculate the recent trend. Years
included in the recent trend are \code{ref_year - num_years}, inclusive}

\item{num_recent_years}{Number of years before \code{ref_year} to include in recent trend}
}
\value{
A data frame with the recent trend calculated by \code{Segment_ID} and \code{ref_year}.
If \code{ref_year} is a vector of years, then all recent trend data is collected into a single
dataframe in long format with columns for \code{Segment_ID}, \code{year}, \code{cumulative_coverage},
and \code{delta}
}
\description{
Calculate recent trends from annual kelp area time series
}
