% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cumulative_dist.R
\name{plot_cumulative_dist}
\alias{plot_cumulative_dist}
\title{Plot ecdf for kelp segments}
\usage{
plot_cumulative_dist(
  annual_time_series,
  kelp_presence,
  ...,
  years = unique(annual_time_series$year),
  ylim = c(0, 200)
)
}
\arguments{
\item{annual_time_series}{Annual time series generated by \code{\link[=extract_time_series]{extract_time_series()}}.
Does not work with quarterly time series}

\item{kelp_presence}{Kelp presence data frame generated by \code{\link[=get_kelp_presence]{get_kelp_presence()}}}

\item{...}{These dots are for future extensions and must be empty.}

\item{years}{Which years to plot. Default is all years present in \code{annual_time_series}}

\item{ylim}{Restrictions on the y axis (\% of historical median). For visual clarity
only. Does not affect any calculations}
}
\value{
ggplot object of the cumulative distributions for each year of kelp segments
}
\description{
Plot ecdf for kelp segments
}
