% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_time_series.R
\name{plot_time_series}
\alias{plot_time_series}
\title{Plot the given kelp area time series}
\usage{
plot_time_series(
  kelp_area_time_series,
  ...,
  type = c("absolute", "percent", "historical"),
  segment_id = NULL
)
}
\arguments{
\item{kelp_area_time_series}{Data frame generated by \code{\link[=extract_time_series]{extract_time_series()}}}

\item{...}{additional aesthetics passed to the ggplot call}

\item{type}{The type of representation for the area value. One of:
\itemize{
\item "absolute": absolute value of the kelp area, in km^2
\item "percent": kelp area relative to the maximum occupiable area, expressed as a percentage
\item "historical": kelp area relative to the historical median, expressed as a percentage
}}

\item{segment_id}{The segment ID(s) to plot. The default option \code{segment_id = NULL} will plot
the first segment in the time series}
}
\value{
ggplot object for the time series
}
\description{
Plot the given kelp area time series
}
