% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segment_landsat_data.R
\name{segment_landsat_data}
\alias{segment_landsat_data}
\title{Assign segment ID's to each pixel in the landsat data according to a
given shapefile}
\usage{
segment_landsat_data(
  lter_file_path,
  kelp_segments_file_path,
  ...,
  segment_id = "all"
)
}
\arguments{
\item{lter_file_path}{character string path to the landsat netCDF file}

\item{kelp_segments_file_path}{character string path to the kelp segments shapefile}

\item{...}{unused}

\item{segment_id}{either "all" or a character vector containing the subset of
\code{segment_id}'s of interest, e.g. \code{c("CA_71", "CA_78")}}
}
\value{
A dataframe with rows for each pixel, with the following columns:
\itemize{
\item \code{Segment_ID}: The ID of the kelp area segment
\item \code{lon}: The longitude of the Landsat pixel
\item \code{lat}: The latitude of the Landsat pixel
\item \code{Q1.1984}: The kelp area in m^2 of the given pixel for quarter 1 of 1984
\item ... : subsequent quarters and years
}

Segments that do not have any pixels ever are represented as rows of \code{NA} values
}
\description{
This function assigns to each Landsat pixel segment ID's derived from a shapefile of polygons with a
Segment_ID attribute. This shapefile describes segments of interest along the west coast of the US.
This function assumes that the Landsat data starts with quarter 1 of the first year of interest.
}
