% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_time_series.R
\name{trend_options}
\alias{trend_options}
\title{Trend plotting options}
\usage{
trend_options(..., trend = NULL, only_recent = FALSE, num_recent_years = 10)
}
\arguments{
\item{...}{Named parameters to pass to the aesthetics of the \code{ggplot2} layer that controls the
trend visualization. In particular:
\itemize{
\item alpha
\item color
\item group
\item linetype
\item linewidth
}}

\item{trend}{Which type of trend to add. One of \code{NULL}, "linear", or "median".
\itemize{
\item \code{NULL}: no trend is plotted
\item "linear": plots a linear regression
\item "median": plots median values of the historical baseline and/or "recent" years
}}

\item{only_recent}{Whether to plot the trend only for the "recent" time period, defined by \code{num_recent_years}}

\item{num_recent_years}{The number of years before the latest year to consider "recent"}
}
\value{
object of class \code{KelpAreaIndicator_trend_options} for use with \code{\link[=plot_time_series]{plot_time_series()}}
}
\description{
Use these options to add/modify a \code{ggplot2} layer to \code{\link[=plot_time_series]{plot_time_series()}},
representing the trend of the time series.
}
